requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
launchdarkly-server-sdk==8.1.1
python-dotenv==1.0.0
